+++ src/main/java/my/sheshenya/samplenettyrestfulapi/repository/AccountRepository.java
@@ -2,14 +2,12 @@ package my.sheshenya.samplenettyrestfulapi.repository;
 
 import my.sheshenya.samplenettyrestfulapi.model.Account;
 
-import java.util.Collection;
-import java.util.List;
-import java.util.Set;
+import java.math.BigDecimal;
 
 public interface AccountRepository {
     Account getAccountById(String accountId);
 
     boolean add(Account account);
 
-    double getTotalBalance();
+    BigDecimal getTotalBalance();
 }
